// sources/Core.React/web-app/src/lib/enums/VocabularyDirectionEnum.ts
export enum VocabularyDirectionEnum {
  PlaceHolder = -1,
  DanishEnglish = 0,
  DanishFrench = 1,
  DanishGerman = 2,
  EnglishDanish = 3,
  EnglishFrench = 4,
  EnglishGerman = 5,
  FrenchDanish = 6,
  FrenchEnglish = 7,
  FrenchGerman = 8,
  GermanDanish = 9,
  GermanEnglish = 10,
  GermanFrench = 11,
}